Move roi_shape and slice_in_out into _tools.py
